backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.1.31
chardet==5.2.0
click==8.1.8
cssselect==1.3.0
fastapi==0.115.12
greenlet==3.2.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
lxml==5.3.2
lxml_html_clean==0.4.2
playwright==1.51.0
pydantic==2.11.3
pydantic_core==2.33.1
pyee==12.1.1
python-dotenv==1.1.0
readability-lxml==0.8.1
sniffio==1.3.1
soupsieve==2.6
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.1
